Simplify ontology builder boundary
