--- packages/core/project-ontology.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/project-ontology.js';
