--- packages/core/project-ontology.js
@@ -1,1 +1,0 @@
-export * from './dist/project-ontology.js';
