+++ packages/core/src/sparql-safe.ts
@@ -15,17 +15,9 @@
  *   Use `sparqlInt`.
  */
 
-import {
-  escapeSparqlLiteral,
-  isSafeIri,
-} from './sparql-safe-runtime.js';
-
-export {
-  escapeSparqlLiteral,
-  isSafeIri,
-} from './sparql-safe-runtime.js';
-
 const UNSAFE_IRI_CHARS = /[<>"{}|\\^`\x00-\x20]/;
+
+const IRI_SCHEME_RE = /^[a-zA-Z][a-zA-Z0-9+.-]*:[^\s<>"{}|\\^`\x00-\x20]+$/;
 
 /**
  * Throws if `value` contains characters that would break a SPARQL IRI (`<...>`).
@@ -41,12 +33,35 @@ export function assertSafeIri(value: string): string {
 }
 
 /**
+ * Returns true when the string is a syntactically safe IRI with a scheme
+ * prefix (e.g. `did:dkg:...`, `http://...`, `urn:...`).
+ */
+export function isSafeIri(value: string): boolean {
+  if (!value) return false;
+  return IRI_SCHEME_RE.test(value);
+}
+
+/**
  * Returns `<value>` after validating the IRI is safe for SPARQL interpolation.
  * Throws on unsafe input.
  */
 export function sparqlIri(value: string): string {
   assertSafeIri(value);
   return `<${value}>`;
+}
+
+/**
+ * Escapes a raw string for use inside a SPARQL `"..."` literal.
+ * Handles all characters that the SPARQL grammar requires escaping
+ * in short string literals (production rule [157] STRING_LITERAL2).
+ */
+export function escapeSparqlLiteral(value: string): string {
+  return value
+    .replace(/\\/g, '\\\\')
+    .replace(/"/g, '\\"')
+    .replace(/\n/g, '\\n')
+    .replace(/\r/g, '\\r')
+    .replace(/\t/g, '\\t');
 }
 
 /**
